Refactor SacDmDevice component to improve code structure and maintainability
