+++ src/pages/Device/index.tsx
@@ -2,7 +2,7 @@ import { useParams } from "react-router-dom";
 import { useEffect, useState, useCallback } from "react";
 import { SacDmProps } from "../SacDm/types";
 import { Description } from "./styles";
-import { SacDmDevice } from "./components/SacDmDevice";
+import { SacDmDevice } from "../SacDm/components/SacDmDevice";
 import { formatTime } from "../../utils/formatTime";
 import { BackPage } from "../../components/BackPage";
 import { AirplanemodeActive } from "@mui/icons-material";
